some fixes

+++ src/main/java/nesemulator/utils/ByteUtilities.java
@@ -2,8 +2,8 @@ package nesemulator.utils;
 
 public class ByteUtilities {
 
-    public static short widenIgnoreSigning(byte byteToWiden) {
-        return (short) ((short) byteToWiden & 0xFF);
+    public static int widenIgnoreSigning(byte byteToWiden) {
+        return ((int) byteToWiden & 0xFF);
     }
 
     public static int widenIgnoreSigning(short shortToWiden) {
